Initial commit.
This is the first iteration of a system to pull wallpapers from the
wallhaven.cc website. At the moment, it only pulls out of the random
category, which means that the wallpapers don't include any NSFW
content. The obvious TODO would be to permit a user to provide
credentials, and have the app log in to the site, and then provide
category tags for returned images.